Big changes.. Extension code refactored from YAML into separate python files.  This was done because handling numpy to num.cr translation (esp with tensor datatypes) has a lot of special cases.  This approach provides mode flexibility at the expense of more code and potentially more verbosity.  The old yamls are still in the repo for posterity and in case anything was missed.
In doing the above changes most of the `from x import y` translation
features broke, (some mightve never worked) and it would be useful to
try to fix those in the future.

Also pulled in many of the py2many tests which dug up some things
that shouldve previously worked but didnt..  others are complex cases
which may never work but perhaps some could be stretch goals?

About 1/3rd of the numpy->num.cr testcases are working, Others are
maybe a smidge away from working.